+++ src/core/complex_vectorised_matrix.h
@@ -17,29 +17,30 @@
 
 #include "complex_vector_split.h"
 #include "hilbert_namespace.h"
+#include <cstddef>
 #include <functional>
 
 class ComplexVectMatrix final {
 public:
   ComplexVectMatrix(const ComplexMatrix &m)
-      : row_size_(static_cast<int>(m.size())),
-        column_size_(static_cast<int>(m.front().size())),
+      : row_size_(m.size()), column_size_(m.front().size()),
         vectorised_matrix_(ComplexVector(row_size_ * column_size_)) {
     const auto elements_size = row_size_ * column_size_;
-    for (int i = 0; i < elements_size; i++) {
+    for (size_t i = 0; i < elements_size; i++) {
       vectorised_matrix_[i] = m.at(i / column_size_).at(i % column_size_);
     }
   }
 
-  ComplexVectMatrix(ComplexVector m, const int row_size, const int column_size)
+  ComplexVectMatrix(ComplexVector m, const size_t row_size,
+                    const size_t column_size)
       : row_size_(row_size), column_size_(column_size),
         vectorised_matrix_(std::move(m)) {}
 
   ComplexVectMatrix(const ComplexVector &v)
-      : row_size_(1), column_size_(static_cast<int>(v.size())),
+      : row_size_(1), column_size_(v.size()),
         vectorised_matrix_(ComplexVector(row_size_ * column_size_)) {
     const auto elements_size = row_size_ * column_size_;
-    for (int i = 0; i < elements_size; i++) {
+    for (size_t i = 0; i < elements_size; i++) {
       vectorised_matrix_[i] = v.at(i);
     }
   }
@@ -52,15 +53,15 @@ public:
   ComplexVectMatrix()
       : row_size_(0), column_size_(0), vectorised_matrix_(ComplexVector(0)) {}
 
-  [[nodiscard]] Complex get(const int m, const int n) const {
+  [[nodiscard]] Complex get(const size_t m, const size_t n) const {
     return vectorised_matrix_.at(m * column_size_ + n);
   }
 
   [[nodiscard]] ComplexVectSplit
-  get(int row_size, std::function<int(int i)> m_functor,
-      std::function<int(int i)> n_functor) const {
+  get(size_t row_size, std::function<size_t(size_t i)> m_functor,
+      std::function<size_t(size_t i)> n_functor) const {
     ComplexVectSplit result;
-    for (int i = 0; i < row_size; i++) {
+    for (size_t i = 0; i < row_size; i++) {
       auto m = m_functor(i);
       auto n = n_functor(i);
       auto index = m * column_size_ + n;
@@ -69,16 +70,16 @@ public:
     return result;
   }
 
-  [[nodiscard]] ComplexVectSplit get_row(const int row) const {
+  [[nodiscard]] ComplexVectSplit get_row(const size_t row) const {
     auto start_index = row * column_size_;
     auto end_index = start_index + column_size_;
     return std::vector<Complex>(vectorised_matrix_.begin() + start_index,
                                 vectorised_matrix_.begin() + end_index);
   }
 
-  [[nodiscard]] ComplexVectSplit get_column(const int column) const {
+  [[nodiscard]] ComplexVectSplit get_column(const size_t column) const {
     ComplexVectSplit result;
-    for (int i = column; i < column_size_; i += column_size_) {
+    for (size_t i = column; i < column_size_; i += column_size_) {
       result.add(vectorised_matrix_[i]);
     }
     return result;
@@ -92,13 +93,13 @@ public:
     return result;
   }
 
-  [[nodiscard]] int row_size() const { return row_size_; }
+  [[nodiscard]] size_t row_size() const { return row_size_; }
 
-  [[nodiscard]] int column_size() const { return column_size_; }
+  [[nodiscard]] size_t column_size() const { return column_size_; }
 
 private:
-  const int row_size_;
-  const int column_size_;
+  const size_t row_size_;
+  const size_t column_size_;
   ComplexVector vectorised_matrix_;
 };
